+++ src/main/java/survivalblock/volucraft/client/datagen/VolucraftRecipeGenerators.java
@@ -22,6 +22,7 @@ import net.minecraft.core.HolderLookup;
 import net.minecraft.data.recipes.RecipeCategory;
 import net.minecraft.data.recipes.RecipeOutput;
 import net.minecraft.data.recipes.RecipeProvider;
+import net.minecraft.resources.Identifier;
 import net.minecraft.tags.ItemTags;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.Items;
@@ -167,6 +168,11 @@ public class VolucraftRecipeGenerators {
         @Override
         public String getName() {
             return "Example Recipes";
+        }
+
+        @Override
+        protected Identifier getRecipeIdentifier(Identifier identifier) {
+            return identifier;
         }
     }
 }