+++ src/main/java/survivalblock/volucraft/common/menu/TransientAmalgamationContainer.java
@@ -15,13 +15,14 @@
  */
 package survivalblock.volucraft.common.menu;
 
-import java.util.List;
 import net.minecraft.core.NonNullList;
 import net.minecraft.world.ContainerHelper;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.entity.player.StackedItemContents;
 import net.minecraft.world.inventory.AbstractContainerMenu;
 import net.minecraft.world.item.ItemStack;
+
+import java.util.List;
 
 /**
  * @see net.minecraft.world.inventory.TransientCraftingContainer