+++ src/main/java/survivalblock/volucraft/common/recipe/datagen/ShapedAmalgamationRecipeBuilder.java
@@ -15,10 +15,6 @@
  */
 package survivalblock.volucraft.common.recipe.datagen;
 
-import java.util.ArrayList;
-import java.util.LinkedHashMap;
-import java.util.List;
-import java.util.Map;
 import net.minecraft.advancements.Criterion;
 import net.minecraft.core.HolderGetter;
 import net.minecraft.data.recipes.RecipeBuilder;
@@ -35,6 +31,11 @@ import org.jetbrains.annotations.ApiStatus;
 import org.jspecify.annotations.Nullable;
 import survivalblock.volucraft.common.recipe.specific.ShapedAmalgamationRecipe;
 import survivalblock.volucraft.common.recipe.specific.ShapedAmalgamationRecipePattern;
+
+import java.util.ArrayList;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
 
 import static net.minecraft.world.item.crafting.ShapedRecipePattern.EMPTY_SLOT;