+++ src/main/java/survivalblock/volucraft/common/recipe/display/ShapedAmalgamationRecipeDisplay.java
@@ -18,13 +18,14 @@ package survivalblock.volucraft.common.recipe.display;
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.MapCodec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
-import java.util.List;
 import net.minecraft.network.RegistryFriendlyByteBuf;
 import net.minecraft.network.codec.ByteBufCodecs;
 import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.world.flag.FeatureFlagSet;
 import net.minecraft.world.item.crafting.display.RecipeDisplay;
 import net.minecraft.world.item.crafting.display.SlotDisplay;
+
+import java.util.List;
 
 /**
  * @see net.minecraft.world.item.crafting.display.ShapedCraftingRecipeDisplay