+++ src/main/java/survivalblock/volucraft/common/recipe/specific/NormalAmalgamationRecipe.java
@@ -15,7 +15,9 @@
  */
 package survivalblock.volucraft.common.recipe.specific;
 
-import net.minecraft.world.item.crafting.*;
+import net.minecraft.world.item.crafting.PlacementInfo;
+import net.minecraft.world.item.crafting.Recipe;
+import net.minecraft.world.item.crafting.RecipeSerializer;
 import org.jspecify.annotations.Nullable;
 import survivalblock.volucraft.common.recipe.AmalgamationRecipe;