+++ changelog.md
@@ -1,2 +1,6 @@
 Volucraft 0.2.5+26.1.2
-- RRV 8.7.0 support
+- RRV 8.9.0 support
+  - Support item filter mode
+- Patch MC-267545 in Amalgamation Table ([bugs.mojang.com](https://bugs.mojang.com/browse/MC/issues/MC-267545) | [mojira.dev](https://mojira.dev/MC-267545))
+- Bump registrar version to 0.1.4
+- Refactor CubeOfSlotsRenderer to not use `ItemStack`s during actual rendering